-Changing the way the camera moves to new location -Adding the path done by currentUser

+++ app/src/main/java/pt/ubi/eventtrackingapp/UserLocationPositionsInEvent.java
@@ -9,7 +9,7 @@ public class UserLocationPositionsInEvent {
     private GeoPoint lastPosition;
 
 
-    public UserLocationPositionsInEvent(String userPositionKey, Integer distanceDone, GeoPoint lastPosition) {
+    public UserLocationPositionsInEvent(String userPositionKey, float distanceDone, GeoPoint lastPosition) {
         this.userPositionKey = userPositionKey;
 
         this.distanceTraveled = distanceDone;
